==> src/notebooks/02_immune_data_combine.ipynb <==
# Putting together chunks for preprocessing
1. Use ```glob``` to find all the chunks
2. Run this pipeline for multidataset integration: https://chanzuckerberg.github.io/cellxgene-census/notebooks/analysis_demo/comp_bio_data_integration_scvi.html
2. Run PCA to get ```X_pca``` layer of the data

# Steps prior to generating the UMAP
1. Normalization
- The CZI data normalises to a sum of 1e1, but the integration procedure uses 1e4. Since there are raw counts available in the data, going ahead with the target sum 1e4 makes more sense. 
2. log1p transformation
3. Scaling



import cellxgene_census
import pandas as pd
import scanpy as sc
import anndata
import numpy as np
import os
from IPython.display import display # easy viewing of dataframes
import glob
from pathlib import Path
import harmonypy
import matplotlib as plt
from scipy.sparse import csr_matrix
import scvi 

# chunk_datapath = Path("../data/cellxgene_data") # set datapath
# chunk_files = glob.glob(str(chunk_datapath) + "/*.h5ad")  # use glob to find all the .h5ad files

# # Read all files into a list
# adata_list = [sc.read_h5ad(f) for f in chunk_files] 
# adata = anndata.concat(adata_list, merge="same") # concatenate adata

# # Set new column batch based on dataset ID
# adata.obs["batch"] = adata.obs["dataset_id"].astype(str)  # Ensure it's a string

# adata.write("../data/datasets/cellxgene_covid_pbmc.h5ad")
# # Restart kernel at this point

# Check for batch induced clustering
covid_data_read = sc.read_h5ad("../data/datasets/cellxgene_covid_pbmc_processed.h5ad")
# del covid_data_read.obsm["X_pca"]

covid_data_read

covid_data_read.obs.cell_type.value_counts()

obs_cols = list(covid_data_read.obs.columns)
print(obs_cols)

for colname in obs_cols:
    if pd.api.types.is_categorical_dtype(covid_data_read.obs[colname]):
        covid_data_read.obs[colname] = covid_data_read.obs[colname].cat.remove_unused_categories()

covid_data_read.obs.dataset_id.value_counts()

covid_data_read.write_h5ad("../data/datasets/cellxgene_covid_pbmc_processed.h5ad")

covid_data_processed = sc.read_h5ad("../data/datasets/cellxgene_covid_pbmc_processed.h5ad")

covid_data_processed.obs.set_index("soma_joinid", inplace=True) # indices are not unique, because of data concatenation

covid_data_processed.obs_names = covid_data_processed.obs_names.astype(str) # convert to str?

czi_covid_pbmc_5pct = sc.pp.subsample(covid_data_processed, fraction=0.05, copy=True, random_state=42)

czi_covid_pbmc_2pct = sc.pp.subsample(covid_data_processed, fraction=0.02, copy=True, random_state=21)

common_obs = czi_covid_pbmc_2pct.obs_names.intersection(czi_covid_pbmc_5pct.obs_names)
print(len(common_obs)) # 4084

czi_covid_pbmc_5pct.write_h5ad("../data/datasets/czi_covid_pbmc_5pct.h5ad")
czi_covid_pbmc_2pct.write_h5ad("../data/datasets/czi_covid_pbmc_2pct.h5ad")

import scanpy as sc
adata = sc.read_h5ad("../data/datasets/czi_covid_pbmc_2pct.h5ad")

adata.var
